requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]
email-validator
pydantic[email]
sqlalchemy
